+++ Flow/Changes.md
@@ -25,6 +25,7 @@
 ### Bugfixes
 - **JsonSettingsAccess** settings file search algorithm was bugged, now fixed
 - Fixed a bug in **Path**`.parts` and `.partsIterator` (used to throw an **IllegalArgumentException**)
+- `ResettableFlag(initialValue = true)` now correctly sets the initial value
 - **StringFrom** now correctly preserves line-breaks
 - **EqualsBy** now compares the equals-properties using `==` instead of using hashCodes
 - **Volatile**`.mutate(...)` now properly locks the wrapped item (whoops!)